% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afromm.R
\name{AfromM}
\alias{AfromM}
\title{AfromM}
\usage{
AfromM(M)
}
\arguments{
\item{M}{A list of square matrices "M" with the Identity as its first matrix}
}
\value{
The Cayley table "A" of the Matrix group "M"
}
\description{
AfromM
}
