% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afromm2.R
\name{AfromM2}
\alias{AfromM2}
\title{AfromM2}
\usage{
AfromM2(M, silent = T)
}
\arguments{
\item{M, silent}{A list of square matrices "M" with the Identity as its first matrix and an option}
}
\value{
The Cayley table "A" of the Matrix group "M"
}
\description{
AfromM2
}
