% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugates.R
\name{Conjugates}
\alias{Conjugates}
\title{Conjugates}
\usage{
Conjugates(A, I, w)
}
\arguments{
\item{A, I, w}{The Cayley table "A" of a group, the list of inverses "I", and a sequence of elements "w".}
}
\value{
A matrix whose rows are the different collections of elements in "A" conjugate to "w".
}
\description{
Conjugates
}
