% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efromax.R
\name{EfromAX}
\alias{EfromAX}
\title{EfromAX}
\usage{
EfromAX(A, X, silent = T)
}
\arguments{
\item{A, X, silent}{The Cayley table "A" of a group and its list of up-to-conjugacy-subgroups "X" and a parameter.}
}
\value{
The adjacency matrix of its up-to-conjugacy-subgroups, where i is linked to j if there is a subgroup of conjugacy-type i that is an immediate subgroup of some subgroup of conjugacy-type j. By immediate subgroup we mean that there is no intermediate subgroup in between.
}
\description{
EfromAX
}
