% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findrows.R
\name{FindRows}
\alias{FindRows}
\title{FindRows}
\usage{
FindRows(v, M)
}
\arguments{
\item{v, M}{A row vector "v" and a Matrix "M".}
}
\value{
The indices of all the rows of "M" approximately equal to "v". The comparison is made only where "M" is non-negative.
}
\description{
FindRows
}
\examples{
FindRows(c(2,3,5),rbind(c(1,2,7),c(2,3,5),c(2,3,-10),c(7,8,5)))

# Example with your dataset in "inst/"
datafile=system.file("M64.RData",package="grouptheory")
load(datafile)
# Apply my function
FindRows(c(1,2,4,8,16,32,7,1,1),M64)
}
