% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupid.R
\name{GroupId}
\alias{GroupId}
\title{GroupId}
\usage{
GroupId(A, w, rk)
}
\arguments{
\item{A, w, rk}{The Cayley table "A" of a group, a vector "w" of elements of a subgroup and the rank of this subgroup.}
}
\value{
the group id
}
\description{
GroupId
}
