% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouptheory-package.R
\docType{package}
\name{grouptheory-package}
\alias{grouptheory}
\alias{grouptheory-package}
\title{grouptheory: Basic Group Theory}
\description{
Identify Groups and work with them.
}
\author{
\strong{Maintainer}: Andrea Aveni \email{andrea.aveni@duke.edu}

}
\keyword{internal}
