% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.R
\name{Inverse}
\alias{Inverse}
\title{Inverse}
\usage{
Inverse(A)
}
\arguments{
\item{A}{The Cayley table "A" of a group}
}
\value{
The vector with the inverse of all the elements in the group.
}
\description{
Inverse
}
