% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is2number.R
\name{Is2Number}
\alias{Is2Number}
\title{Is2Number}
\usage{
Is2Number(x)
}
\arguments{
\item{x}{A natural number}
}
\value{
1 or 2 if there are exactly two groups of order x, 0 otherwise.
}
\description{
Is2Number
}
