% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesfromy.R
\name{NamesfromY}
\alias{NamesfromY}
\title{YNamesfromY}
\usage{
NamesfromY(Y)
}
\arguments{
\item{Y}{The matrix group "M", its Cayley table "A", and its list of up-to-conjugacy-subgroups "X".}
}
\value{
The names of the groups
}
\description{
YNamesfromY
}
