% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberofsubgroups.R
\name{NumberOfSubgroups}
\alias{NumberOfSubgroups}
\title{NumberOfSubgroups}
\usage{
NumberOfSubgroups(A)
}
\arguments{
\item{A}{The Cayley table "A" of a group}
}
\value{
A couple with the number of subgroups up to conjugacy and overall.
}
\description{
NumberOfSubgroups
}
