% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences.R
\name{Occurrences}
\alias{Occurrences}
\title{Occurrences}
\usage{
Occurrences(v, M)
}
\arguments{
\item{v, M}{A matrix "M" and a row vector "v".}
}
\value{
The number of times "v" is a row vector of "M".
}
\description{
Occurrences
}
\examples{
Occurrences(c(2,3),rbind(c(1,2),c(2,3),c(2,3),c(7,8)))

# Example with your dataset in "inst/"
datafile=system.file("M64.RData",package="grouptheory")
load(datafile)
# Apply my function
Occurrences(c(1,2,4,8,16,32,7,1,1),M64)
}
