% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderstructure.R
\name{OrderStructure}
\alias{OrderStructure}
\title{OrderStructure}
\usage{
OrderStructure(A)
}
\arguments{
\item{A}{The Cayley table "A" of a group.}
}
\value{
The number of elements of each possible orders except one. That is of each non-one divisor of the size of the group.
}
\description{
OrderStructure
}
