% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{Relabel}
\alias{Relabel}
\title{Relabel}
\usage{
Relabel(A, w)
}
\arguments{
\item{A, w}{The Cayley table "A" of a group and a vector "w" of elements in a subgroup}
}
\value{
The Cayley table "A" of the subgroup.
}
\description{
Relabel
}
