% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{Span}
\alias{Span}
\title{Span}
\usage{
Span(A, w)
}
\arguments{
\item{A, w}{The Cayley table "A" of a group and a vector "w" of elements in the group}
}
\value{
The vector of elements in the subgroup of "A" generated by "w".
}
\description{
Span
}
