% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfroma.R
\name{XfromA}
\alias{XfromA}
\title{XfromA}
\usage{
XfromA(A, silent = T)
}
\arguments{
\item{A, silent}{The Cayley table "A" of a group and a parameter}
}
\value{
A list of all the subgroups of "A" up to conjugacy. Each element of the list is a vector containing 1. its size, 2. its occurrences, 3. its rank ie the minimal size of a generating set, 4. the vector of the elements of one of its representatives.
}
\description{
XfromA
}
