% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yfrommax.R
\name{YfromMAX}
\alias{YfromMAX}
\title{YfromMAX}
\usage{
YfromMAX(M, A, X)
}
\arguments{
\item{M, A, X}{The matrix group "M", its Cayley table "A", and its list of up-to-conjugacy-subgroups "X".}
}
\value{
A list of identifiers of the groups
}
\description{
YfromMAX
}
